don't kill jobs on every change
This changes the job running strategy.

Previously, the current running job was killed as soon as there was a
change, to allow a faster execution of the new job.

This had the downside of never allowing any job to complete if you were
always changing files and the job were just a little too long to finish
between changes.

With the new strategy, bacon lets the job finish (unless you ask for a
refresh or change the mission), display the result, and if some changes
occured after the last job start, then start a new job.

Also, the number of changes since the last job *start* is displayed
(this display should become opt-in before merging into main)
(note: it's the number of events, not of modified files)

Fix #206