make sound list appendable
To allow custom sounds, we need a mutable sound list. This is achieved
through an once_cell::Lazy<RwLock<HashMap<&str, Sound>>>. It's a global
static because the sound list is basically global in each run. RwLock is
suitable for our situation where writes are rare and reads are often.
HashMap is to ease retrieving sounds by name.

Names and bytes are leaked because they are basically static in each
run.