read sound duration at runtime
With custom sounds, there's no way to know their durations beforehand,
like for default sounds. rodio has the `Source::total_duration()` trait
method, which is implemented on its `Decoder`, so that is used. But it
returns an `Option<Duration>`, i.e. reading is not guaranteed. Thus we
ignore the interrupt for such sounds, for now.