Add Cargo config so devs can take advantage of new resolver
This allows devs that have a newer Rust toolchain to take advantage of
features in edition2024 such as the MSRV aware resolver without forcing
editing2024 or a newer MSRV on actual consumers of the program. With
this in place users of new toolchains can run `cargo update` and the
updates will only bump as far as the MSVR allows. Use `cargo update -v`
for information about what updates were held back and why.