Update lock file respecting MSRV
This resets all the dependencies in the lock file to a lockfile actually
generated with an MSRV aware resolver (just using `cargo update` from
a new toolchain but with the Cargo config file in place). This will
allow the app to build using the lock file on the advertised minimum
supported rust toolchain.